add router and component home
