+++ mdswac/src/components/test.css
@@ -1,0 +1,40 @@
+.homepage{
+    margin-top:20vh;
+    margin-left:20vh;
+    background-image: url(https://zupimages.net/up/19/51/x8ma.jpg);
+    width:70%;
+    height:60vh;
+    display:flex;
+    flex-direction:column;
+}
+
+.EnterTitle{
+    width:70%;
+    height:50vh;
+    background-color: whitesmoke;
+    margin-bottom: 15%;
+    margin-top: 15%;
+    margin-left: 15%;
+    margin-right: 15%;
+    opacity:0.8;
+    text-align:center;
+    vertical-align:center;
+    align-items: center;
+}
+.EnterTitle p{
+    margin-top: 10%;
+    font-weight: 900;
+    font-size:1.5em;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+}
+
+.EnterButton {
+    margin-left: 42%;
+    margin-top: 5%;
+    border:2px solid white;
+    color:black;
+    box-shadow: 8px 12px 8px rgb(51, 50, 50);
+    font-weight: 900;
+    font-size:1.1em;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+}
